+++ src/hklpy2/backends/th_tth_q.py
@@ -20,7 +20,7 @@ from typing import List
 
 from .. import __version__
 from ..misc import IDENTITY_MATRIX_3X3
-from ..misc import SolverError
+from ..exceptions import SolverError
 from ..typing import Matrix3x3
 from ..typing import NamedFloatDict
 from .base import SolverBase